Rewrite simple sound test to an Application. Fix AudioManager's rewind feature and cleaning the sources
