+++ archimedes_bin/main.cpp
@@ -5,7 +5,7 @@
 int main() {
 	arch::Logger::init(arch::LogLevel::trace);
 
-	arch::Ref<arch::Application> myApp = arch::createRef<SpatialAudioTestApp>();
+	arch::Ref<arch::Application> myApp = arch::createRef<SimpleSoundTestApp>();
 
 	arch::EngineConfig config{ .windowWidth = 1'200,
 							   .windowHeight = 600,
